requirements.txt
python-dotenv==1.0.1
pyTelegramBotAPI==4.17.0
asyncpg==0.29.0
aiohttp==3.9.5
pytest==8.2.0
pytest-asyncio==0.23.6